Created a commons abstract class for HostapdConfigWriter and WpaSupplicantConfigWriter.
